v2.0-stable/requirements.txt
deepgram-sdk>=4.1.0,<4.2.0
pyaudio>=0.2.13
websockets>=10.4
python-dotenv>=1.0.0